\name{MLvdm-package}
\alias{MLvdm-package}
\alias{MLvdm}
\docType{package}
\title{
\packageTitle{MLvdm}
}
\description{
\packageDescription{MLvdm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MLvdm}
\packageIndices{MLvdm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MLvdm}

Maintainer: \packageMaintainer{MLvdm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
